requirements.txt
django>=3.1.2
django-allauth>=0.42.0
django-recaptcha>=2-1.4.1
django-rest-framework>=0.1.0 
django-widget-tweaks>=1.4.8
kombu>=4.6.8
psycopg2==2.8.5
python-memcached>=1.59